[nwp] SPPT review
See merge request icon/icon-nwp!996

Review SPPT and fix issue leading to the observed t2m and td2m biases (cold and dry). 

In the application of the SPPT tracer perturbations a multiplication of the perturbations (tendencies in unit/s) with timestep was missing.

The branch also includes the commits of the branch sppt_opposed_member_perturbation from Daniel Hupp: use for every second member a new random pattern and for the others the opposed pattern from the previous member. On paper it provides a better sampling of the uncertainty for small  ensembles. Tests showed, however, no impact on the observed biases.Surprisingly, they disappeared with the fix for the perturbations of the tracer tendencies.

Approved-by: Pavel Khain <pavelkh_il@yahoo.com>
Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>