[nwp] Reorganization of nwp buildbot data
See merge request icon/icon-nwp!1065

On `rcl.dwd.de` the NWP-specific buildbot checksuite data have been moved 

from
```
/hpc/rwork0/routfor/test/icon/Checksuite_data
```
to
```
/hpc/rwork0/routfor/test/icon/buildbot_data/nwp/Checksuite_data
```

This matches with the current directory structure on `DAINT`, `BALFRIN` and `LEVANTE` and provides the basis for cross-system execution of the existing experiments. For backward compatibility, a symbolic link 

```
/hpc/rwork0/routfor/test/icon/Checksuite_data -> /hpc/rwork0/routfor/test/icon/buildbot_data/nwp/Checksuite_data
```
has been created.

This branch contains necessary adaptions to the experiment files, in order to take the new directory structure into account.


@b380572

Approved-by: Marek Jacob <marek.jacob@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>