Bug fix for soil water check in initicon
See merge request icon/icon-nwp!1069

This branch fixes a bug in the limitation of the soil water content in initicon, which led to a crash in a NUMEX experiment. So far, the limitation of the soil moisture after adding the SMA increments used the non-tiled soil type, which works correctly except for partial glacier points, for which the soil type is adjusted in order to be consistent with the corresponding landuse class. In addition, the on-demand debugging output in TERRA to investigate numerical instabilities is restored in order to be useful for ICON.

Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>