[nwp] Remove optional divergence averaging
See merge request icon/icon-nwp!992

We remove the option for bilinear averaging of horizontal divergence (`idiv_method=2`). \
At an early stage of the dynamical core development, averaging of the divergence was used in order to obtain a nearly second-order accurate divergence operator and to control the checkerboard noise emerging from the triangular grid. The divergence averaging has been discarded at some point in time and nowadays the standard divergence operator (Gauss integral) is applied to **horizontally averaged** normal velocity components. This has proven to perform at least equally well in terms of noise damping, and does not compromise local mass conservation.

The removal of `idiv_method=2` allowed for a significant simplification of the subroutine `prepare_tracer (mo_prepadv_util)`. 
Now it is only used in standalone advection runs for updating the `prep_adv` state for tracer transport. It is no longer used for realcase runs, as the update of the `prep_adv` state now happens unconditionally inside the dynamical core.

The modifications to the subroutine `prepare_tracer` have been checked for correctness by comparing results of the DCMIP Hadley test case against a reference from the NWP master. Results are bit-identical.

### Reproducibility issue on levante_intel

After the removal of the switch `idiv_method=2` from the dynamical core, the experiments `check.atm_ape.run` and `check.atm_bubble_test.run` failed on `levante_intel`.
https://bb-prod8.dkrz.de/#/builders/21/builds/2902

The ape experiment failed the `nproma` test, while the bubble test fails the `mpi` test. The reason for non-reproducible results remained unclear, but we were able to work around by disabling the FMA compiler option for the dynamical core, as shown in 
https://bb-prod8.dkrz.de/#/builders/21/builds/2918

It turned out that the performance penalty of disabling FMA for the dynamical core on levante seems to be insignificant (if at all present).

Some manual timings for `check_atm_ape`

| timer | dycore with FMA | dycore without FMA|
| ------| ------          | ------ |
| nh_solve     | 1.864s | 1.806s |
| integrate_nh |   56.802s     |   56.732s    |


The reproducibility problem disappeared again, once the changes related to the deep atmosphere implementation entered the dynamical core (icon-nwp!1014). Hence, we have been able to switch on FMA for the dynamical core again.


@k202125

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>