[nwp] Modified z_pbl output for icon-nwp
See merge request icon/icon-nwp!683

This merge request is initiated for the implementation of the output variable boundary layer height (z_pbl) in the ICON-NWP code. Before, the calculation of z_pbl in ICON was only possible for LES simulations. In order to enable a calculation for standard ICON simulations, the code was transferred to another section in the routine. It is only calculated when the variable is declared in the namelists. In order to distinguish between the calculation in LES-mode and in standard (non-LES-) mode, the variable in standard mode has been naed "hpbl".
 
Introduced was a slight correction of the calculation itself. Before only the absolute height was written out, now it is the height above ground. This way also a failure in areas with ground below zero is fixed. 
In contrast to COSMO code, no differentiation between different stability conditions, i.e. stable/unstable/neutral is made. This would be the next step for more realistic results. 

It should also be noted, that it is still not possible, to calculate the hpbl in standard mode on a GPU-HPC. For that, the code has to be ported to GPU first.

Code has been changed in:

src/atm_phy_nwp/mo_nwp_diagnosis.f90 

src/atm_phy_nwp/mo_nwp_phy_state.f90

src/configure_model/mo_io_config.f90

@m300186

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>