[mpim] Add adiabatic source term to thermodynamic equation
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/160

ICON uses theta-v as a thermodynamic coordinate which it updates through a
prognostic equation for the exner function pi. The original implementation
misses the adiabatic source term that arises due to differences in the specific
heats of the varied components of moist air, and which multiply the convergence.
A proper treatment of this term has been included in the dynamical core, by
modifying the non-hydrostatic solver updating of the exner function.

The implementation involves the definition of some helper functions and is
activated through a namelist variable that allows the adiabatic source term to
be neglected, which was desirable for the computational efficiency of the NWP
configurations, where for short runs it does not appear to substantially affect
quantities of interest.

Co-authored-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>
Co-authored-by: Guenther Zaengl <Guenther.Zaengl@dwd.de>
Co-authored-by: Marek Jacob <marek.jacob@dwd.de>
Co-authored-by: Monika Esch <monika.esch@mpimet.mpg.de>
Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>