ACC: Fix barycentric lonlat interpolation
See merge request icon/icon-nwp!1034

The old update of the coefficients was too early in the setup phase.

Now the transfer of the coefficients is centralized and the new code is tested in the exp.dwd_run_ICON_09_R2B4N5_EPS experiment.

In addition, `cover_koe_config%inwp_gscp` was accessed from GPU but never initialized. This is also fixed now by updating all components of `cover_koe_config` on GPU.

Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Merged-by: Florian Prill <florian.prill@dwd.de>