CAMS climatological aerosol as an option for ecRad radiation
See merge request icon/icon-nwp!783

The new 3D CAMS climatology is introduced (https://www.ecmwf.int/en/elibrary/80028-implementation-cams-based-aerosol-climatology-ifs) .
In this implementation the original CAMS data is horizontally interpolated using CDO in the pre-processing step.
The vertical interpolation is done inside ICON. There are currently 11 aerosols species and the data is of layer integrated mass. This 3D fields are than transformed to mixing ratios using the gravitational constant divided by pressure thickness. The interactions of these aerosols with radiation are already implemented in ecRAD (including the optical properties).
This implementation does not read the CAMS data from an extpar file as this would mean to keep all 12 months of the 4D data (horizontal, vertical, 12 variables) in memory which would be a lot, especially for high res simulations. Instead, an implementation of the `abstract_reader` is added for CAMS which keeps only the two needed months in the memory. As the abstract reader had been designed for specific time stamps in the data rather than climatological data, some modifications in `mo_interpolate_time` were necessary. As an outlook, this implementation will be extended in a follow-up merge request to be capable of handling CAMS forecasted aerosol as well. For this purpose, this approach of using the abstract reader makes even more sense.

Approved-by: Daniel Rieger <daniel.rieger@dwd.de>
Merged-by: Ulrich Schaettler <ulrich.schaettler@dwd.de>