bc_greenhouse_gases: correct variable names to represent that data from file are VMRs
See merge request icon/icon-nwp!1081

The previous naming stipulated that the values from the bc_greenhouse_gases.nc file
were MMRs, but the values were used as-is in RRTMGP as VMRs. In many other places in
ICON-A (jsbach & ocean coupling), and consistently throughout ICON-NWP, the values are
used as MMRs. This lead to confusion and surprising results when moving from namelist-
specified VMRs to values read from file, which the code expects to be MMRs.

This MR changes the meaning of the values stored in bc_greenhouse_gases.nc to be VMRs
and names the corresponding variables appropriately. MMRs are provided as well for
convenience and to minimize changes to the parts of the code that already expect MMRs
in the MMR variables.

To make the actual GHG VMRs transparent to the user, printouts at simulation start
(msg_level 5) and for each month or day are added at msg_level 5 and 11, respectively.

Merged-by: Daniel Rieger <daniel.rieger@dwd.de>