[mpim] [icon-mpim:feature-if-aes-phy-3] Bug fix for physics update of the vertical wind
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/243

The vertical wind was updated twice in the physics because the pointer
f%wa for vertical wind in physics was set to the prognostic field dyn_new%w
instead to a field that can be used as work field in physics.

Therefore dyn_new%w was updated wrongly in interface_aes_vdf, where f%wa was
updated for later use in other physics computations with the tendency from
the 3d turbulence scheme, and at the end of physics, in interface_iconam_aes,
correctly with the collected physics tendency for the vertical wind, resulting
in a double update.

Here this problem is resolved by pointing f%wa to dyn_now%w instead of
dyn_new%w so that physics has its own memory for wa. dyn_now%w can be used for
this purpose because it is unused after the dycore computations.

The vertical wind is now copied from dyn_new%w to f%wa (that means to dyn_now%w)
and then synchronized before calling the physics main routine.

This bug fix is tested in a modified test experiment exp.checku.atm_qubicc_nc
that uses aes_vdf_config(1)%turb = 2 to activate the 3d turbulence scheme so
that the original bug has effects.

The result with this bug fix is the same as if the wrong update in phyiscs,
in interface_aes_vdf, or the correct update in in the interface is commented
out so that only one update is effective.

Merged-by: Monika Esch <monika.esch@mpimet.mpg.de>