[nwp] Fix intense memory usage on Balfrin
See merge request icon/icon-nwp!1115

There are some dynamically allocated arrays in src/advection, which leads to excessive memory usage when running on GPU. The problem is that OpenACC only de-allocates the dynamic arrays when the memory usage is close to the upper limit. This could trigger an MPI bug in the CRAY MPI library.

According to investigations by @b380976, @b381656, @b382090, and @m300653, we introduce some `ifdef` statements to allocate the arrays differently when compiling with OpenACC. In case of GPU compilation, the arrays are fixed to the size `nproma*p_patch%nlev`. In case of CPU compilation, we keep the regular dynamic allocation, where the arrays are of size `falist%npoints`.

Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>