{atmo,ocean}_coupling_frame: use time_config%tc_dt_model instead of run_config:modelTimeStep
See merge request icon/icon-nwp!1130

The `modelTimeStep` variable just contains the raw string read from run_nml. It is not
set if the time step is set via run_nml:dtime, in which case YAC aborts the simulation.
    
In contrast, `tc_dt_model` is always set to the timedelta corresponding to the timestep
specified by the user, no matter how they specified it.

This MR also removes the modelTimeStep from the `seamless_bb-proto2` experiment in order
provoke the abort in case of a regression.
    
Note: jsbach's HD to ocean interface has the same problem and has to be fixed separately.

Approved-by: Martin Köhler <martin.koehler@dwd.de>
Merged-by: Daniel Rieger <daniel.rieger@dwd.de>