Performance data access
See merge request icon/icon-nwp!1142

the mch-opr list makes performance tests and stores the data. The currently is only readable by the icontest user. This MR fixes that and makes the data automatically readable globally.
Also added itype_conv to the namelist as this is what we use operationally as well.

Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>