[nwp] Coupling with hydrological model HD
See merge request icon/icon-nwp!1105

The hydrological model HD was coupled with ICON-NWV/ICON-O via YAC. HD is added as an external submodule of ICON, so that the model will be easily updated and maintained. This is different from the HD version internally inside Jsbach. In this version, HD is coupled with Jsbach and Terra.

The coupled fields are surface/subsurface runoffs (ICON-NWV) -\> HD; river discharge (HD) -\> ICON-O.

Approved-by: Roland Wirth <roland.wirth@dwd.de>
Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>