[nwp] Move kstart in turbdiff interface to metadata of modvar
See merge request icon/icon-nwp!1114

The 'modvar' structure was introduced as interface for tracers to turbulence. On interface level the corresponding object is called 'ptr'. A big advantage of this structure is that it makes turbulence agnostic with respect to tracer indexing in ICON. This is also very useful for the ComIn tracer implementation.

However, the devil is in the details. At some time, a ' kstart_tracer' was introduced for ART tracers and ' idx_nturb_tracer' was also explicitly passed for it, which massively complicates the Comin implementation. However, I have a fairly simple fix by making kstart a metadata of the modvar structure. This makes it much clearer, and it also brings the implementation back in line with the original concept.

Approved-by: Matthias Raschendorfer <matthias.raschendorfer@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>