[nwp] Horizontal transport of ocean surface wave energy (part I)
See merge request icon/icon-nwp!1165

Implements the horizontal propagation of surface wave energy up to cell row 3.

This merge request exclusively contains changes to the waves code.
It uses parts of the existing horizontal transport scheme for passive and dynamically active tracers from the atmosphere model. 
Any necessary generalizations of the transport code are postponed to part II of this work in icon-nwp!1165.

The following components have been added to the waves code

* [x] new namelist `energy_propagation_nml` for configuring details of the wave energy propagation process.
* [x] new experiment `nwpexp.run_ICON_18_R2B4_waves_adv_nophys`, which solves the energy equation in an idealized setup without forcing terms, i.e. pure energy propagation without physics.
* [x] a new wave-specific transport interface `mo_wave_advection_stepping`, which handles the propagation of wave energy. It uses the horizontal flux routine `upwind_hflux_miura` from the tracer transport module.
* [x] the computation of wave refraction has become part of the transport interface.
* [X] cleanup and fixes:
   - reading of the SLEVE namelist (sleve_nml) has been remved
   - the `model_init` timer has been fixed
   - fix a couple of `INTENT` attributes in the physics.


@b382375

Approved-by: Mikhail Dobrynin <mikhail.dobrynin@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>