[mpim] ocean: support proc0_shift
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/230

On NEC Aurora, separate I/O procs on vector hosts require that PE0 also runs
on a vector host. That is accomplished by setting the namelist parameter
proc0_shift=1, which assigns an empty domain to PE0. The ocean is, however,
not expecting empty domains and segfaults.

This changeset improves the handling of this edge case, allowing ocean
processes to run with an empty domain. The changes comprise some general
improvement of index handling, some fixes to the dbg_print utility (which
now produces sensible output for non-sequential runs), and fixes for the CG
solver. Other solvers are not expected to work without additional
modifications. A clean abort is ensured by a cross-check if proc0_shift is
active.

To ensure that PE0 takes the same code path as the vector engine processes,
LVECTOR is added to the ocean fcgroup flags on the vector host.

This changeset also enables ocean proc0_shift (conditional on being
enabled for the atmosphere) and separate I/O procs for NWP experiments 17
(R2B4_AO_coupled) and 20 (R2B4_R2B6_AO_coupled).

Approved-by: Helmuth Haak <helmuth.haak@mpimet.mpg.de>
Merged-by: Helmuth Haak <helmuth.haak@mpimet.mpg.de>