Fix uninitialized data reads in kenda-ch1_small
See merge request icon/icon-nwp!1085

This MR fixes all uninitialized reads for kenda-ch1_small.
Most of the are most likely harmless. But it is convenient to have them fixed, for investigating future uninitialized memory read.
One of them is a clear bug.
This MR also adds a new BB test to check for uninitialized data reads.

Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Ulrich Schaettler <ulrich.schaettler@dwd.de>