Rewrite gamma functions to avoid license issues
See merge request icon/icon-nwp!1156

Re-write code for the implementation of the gamma functions, so that it is not the code of Numerical Recipes any more:
- ordinary gamma functions (replaced by intrinsic gamma())
- incomplete gamma functions

The replacement of gfct() and gamma_fct() with the intrinsic gamma() causes some slight changes in the 10th significant digit for the gamma function evaluation. Therefore, slight changes of results are to be expected when using either the 1-moment or the 2-moment cloud microphysics.

Also, similar things have been done in submodule emvorado, so that we also expect changes of results in radar reflectivity, if emvorado is used. But this is part of a separate pull request.

Approved-by: Axel Seifert <axel.seifert@dwd.de>
Merged-by: Ulrich Schaettler <ulrich.schaettler@dwd.de>