[nwp] Adapt HZEROCL diagnostic
See merge request icon/icon-nwp!1173

Operations at MeteoSwiss noticed that HZEROCL (the diagnostic for the uppermost freezing level) needs to be adapted. The issue is the case where T < 0 °C on all model levels (meaning no freezing level throughout the atmosphere). Currently, HZEROCL is then set to the orography height. We add a new namelist switch itype_hzerocl to change this. The options are:
- iytpe_hzerocl = 1 ! orography height (DEFAULT),
- itype_hzerocl = 2 ! undefined,
- itype_hzerocl = 3 ! extrapolated height below ground.

We set itype_hzerocl = 2 for all MeteoSwiss experiments that include HZEROCL as an output field.

Approved-by: Fabian Gessler <fabian.gessler@meteoswiss.ch>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>