[nwp] Remove occurrences of iequations from the code
See merge request icon/icon-nwp!1215

The namelist switch `iequations (dynamics_nml)` is used in the atmosphere model and the ocean model to distinguish between different sets of governing equations. Positive integer numbers are used for the atmosphere model, while negative integers are used for the ocean. 

In early days, the following set of governing equations existed for the atmosphere:
* the shallow water equation set
* the hydrostatic equation set (in two variants)
* the nonhydrostatic equation set

Meanwhile, the hydrostatic and shallow water code has been removed from the atmosphere model, with only the nonhydrostatic equation set being left. In the ocean there is also only a single set of governing equations available (hydrostatic ocean).

As a result, `iequations` is no longer required in order to distinguish between different sets of equations. If at all, `iequation` might still be used to distinguish between the atmosphere model and the ocean model. However, the recommended way for identifying the model component is to query the namelist switch `model_type (master_model_nml)`. 

As a result, `iequations` is no longer needed.

#### Additional changes

* the calls in the driver routine `mo_atmo_model:atmo_model` for the nonhydrostatic atmosphere model have been streamlined
* ocean-specific fields have been removed from the interpolation state `t_int_state`. They are already defined in `t_operator_coeff` (`src/ocean/math/mo_ocean_types.f90`)
* the obsolete field `tplane_e_dotprod` has been removed from the interpolation state, including the calculation routine `calculate_dotproduct_at_edge`.

#### Related changes in jsbach

(they will enter the nwp-master via RC)

* iequations cleanup: \
  https://gitlab.dkrz.de/jsbach/jsbach/-/merge_requests/127

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>