[nwp] Add ACM license option
https://gitlab.dkrz.de/icon/icon/-/merge_requests/260

This MR adds a new option to enable / disable code parts under ACM license.  
This affect in particular sppt.  
The option is by default False. The corresponding preprocessor macro in the code is HAVE_ACM_LICENSE.  
For Balfrin and Daint build the option is activate in order to test sppt.
Updated Balfrin and Daint spack verion to v0.20.1.4

Co-authored-by: Xavier Lapillonne <xavier.lapillonne@meteoswiss.ch>
Co-authored-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>
Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Merged-by: Florian Prill <florian.prill@dwd.de>