Revised implementation of moist thermodynamic correction term (lmoist_thdyn)
See merge request icon/icon-nwp!1216

The implementation of the moist thermodynamic correction term (`lmoist_thdyn`, accounting for the moisture-dependence of the air heat capacities) has been revised in order to be consistent with the numerical treatment of moisture tracers. That is, the fluxes are accumulated over the dynamics substeps, and the remaining computations are executed at the fast-physics time step. Given the fact that the moisture correction term is not part of the sound wave equation, there is no need to perform the calculations at the dynamics time step, let alone implicitly. Besides simplifying the code, the revised implementation **reduces the additional computational cost by almost a factor of 5**.

While comparing the revised implementation with the existing one, it became evident that the metric terms arising from the coordinate transformation are not correctly accounted for in the latter. Specifically, the contravariant correction for vertical velocity is fully applied to the explicit part of the computations and missing for the implicit part, leading to inconsistencies over sloping terrain that show up as erroneous temperature changes at and near the surface. These errors disappear when performing the calculations fully explicitly, regardless of whether this is done within the dycore or outside at the fast physics time step, but the latter is clearly preferable for efficiency reasons.

Approved-by: Marek Jacob <marek.jacob@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>