[nwp] Revise construction of tangent planes at edge midpoints for backward trajectory calculations
See merge request icon/icon-nwp!1243

The _flux area_ construction for the tracer transport schemes (Miura, Miura3, FFSL, ...) is performed in a 2D cartesian coordinate system, its origin located at the edge midpoint and the x and y axis pointing in the edge normal and tangent direction.
The flux area construction requires knowledge about the coordinates of the edge end points (vertices) and the center of the cells sharing the given edge. 
These points are projected to the cartesian system with the Gnomonic projection method.

In this merge request, the construction of the tangent plane has been slightly revised. In particular 

* the number of points that are projected to the cartesian system has been reduced. I.e. the edge midpoints of the quadrilateral spanned by the neighbor triangles are no longer needed by the current transport schemes.
* the memory layout of the field `p_int%pos_on_tplane_e` has been changed from `(nproma,nblks_e,npoints,coord)` to `(nproma,npoints,coord,nblks_e)`.
* the routines `calculate_tangent_plane_at_edge` and `calculate_planar_distance_at_edge` that compute `p_int%pos_on_tplane_e` for spherical and planar geometry, respectively, have been streamlined and cleaned.


No change of the results is expected.

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>