Additional ensemble perturbations for deep convection
See merge request icon/icon-nwp!1252

Additional ensemble pertubations for the deep convection part of the Tiedtke-Bechtold scheme are implemented in order to obtain more ensemble spread in convectively active regions. This will allow reducing the SST perturbations in the global ensemble, which currently are unrealistically large, particularly in the tropics.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>