Small bug fix for adaptive surface friction
See merge request icon/icon-nwp!1257

This branch fixes a small bug in the adaptive surface friction that became apparent during the review/testing of Matthias' "mrprep" branch, where this bug happened to be fixed without being aware of it. There is no measurable impact on forecast quality because the bug affects only the diagnosis of the snow-cover fraction on grid points that were fully snow-covered at the previous time step, but the fix is needed to obtain bit-identical results with mrprep using the operational model configuration.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>