Tunable parameter for adaptive surface friction
See merge request icon/icon-nwp!1271

In the global ICON assimilation cycle, the effective weight of 10-m wind observations was found to be significantly smaller in the ensemble data assimilation (LETKF) than in the variational assimilation for the deterministic run. This is somehow related to the climatological part of the B matrix, but it is unclear why the difference is much more pronounced for the 10-m winds than for 2-m temperature and humidity. To partly compensate this difference, the scaling parameter for adaptive surface friction is changed to be tunable. Doubling it for the EPS improves the CRPS for the 10-m winds by 2-3%.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>